modified utilities parsing mlir files for analysis
