+++ utils/IsolatePass.py
@@ -15,9 +15,8 @@
 import argparse
 import os
 import re
-from collections import Counter
 
-from mlir_log_utils import clean_line
+from mlir_log_utils import clean_line, apply_comments, apply_wrap
 
 # Global variables.
 
@@ -79,6 +78,13 @@ def get_args():
         "--comments",
         action="store_true",
         help="Add useful comments to the printed listing: use-count for each def; values for scalar constant.",
+    )
+    parser.add_argument(
+        "-w",
+        "--wrap",
+        action="store_true",
+        help="Reformat long func signatures / ops / returns across multiple indented "
+        "lines when they exceed the line-length limit (same transform as AnalyzeShape.py).",
     )
     return parser.parse_args()
 
@@ -210,202 +216,6 @@ def locate_pass(name, num):
 
 
 ###########################################################
-# Optionally annotate each def by its number of uses
-
-# Matches SSA values like %0, %res_12, %xyz.$tmp
-SSA_RE = re.compile(r"%[A-Za-z0-9_.$]+")
-
-
-def split_lhs_rhs_strip_comment(line: str):
-    """Return (lhs, rhs) after stripping '//' comments.
-    If '=' not present, lhs == '' and rhs == (comment-stripped line).
-    """
-    # Drop '//' comments
-    line = line.split("//", 1)[0]
-    if "=" not in line:
-        return "", line
-    lhs, rhs = line.split("=", 1)
-    return lhs, rhs
-
-
-def split_comment(line: str):
-    """Split a line into (code, comment) where comment includes the leading '//' if present."""
-    if "//" in line:
-        code, comment = line.split("//", 1)
-        return code, "//" + comment
-    return line, ""
-
-
-def collect_outputs(mlir_text: str):
-    """Collect names defined on the LHS of '=' (i.e., operation results)."""
-    outputs = set()
-    for line in mlir_text.splitlines():
-        lhs, _ = split_lhs_rhs_strip_comment(line)
-        if lhs:
-            for name in SSA_RE.findall(lhs):
-                outputs.add(name)
-    return outputs
-
-
-def count_output_uses(mlir_text: str):
-    """Return a dict { '%name': use_count } for every SSA defined as '%x = ...'.
-
-    Counts only RHS occurrences (and lines without '=') so definitions aren't counted.
-    Includes outputs with 0 uses.
-    """
-    outputs = collect_outputs(mlir_text)
-    counts = Counter({name: 0 for name in outputs})
-    for line in mlir_text.splitlines():
-        lhs, rhs = split_lhs_rhs_strip_comment(line)
-        # Count only the RHS (or the whole line if there's no '=')
-        scan = rhs  # rhs already equals full line when lhs == ''
-        for name in SSA_RE.findall(scan):
-            if name in outputs:
-                counts[name] += 1
-    return dict(counts)
-
-
-def annotate_mlir_with_use_counts(mlir_text: str, use_counts: dict[str, int]) -> str:
-    """
-    Return a new MLIR text where each SSA result defined on a line is annotated as '%x/N'
-    on the LHS of '=', where N is the use count from `use_counts`.
-
-    Non-definition lines are left unchanged. Comments are preserved.
-    """
-    out_lines = []
-    for line in mlir_text.splitlines():
-        code, comment = split_comment(line)
-        if "=" not in code:
-            # Not a definition line—pass through unchanged
-            out_lines.append(line)
-            continue
-        lhs, rhs = code.split("=", 1)
-
-        # Replace each SSA name in the LHS with '%name/N'
-        def _add_count(m: re.Match) -> str:
-            name = m.group(0)
-            return f"{name}/{use_counts.get(name, 0)}"
-
-        annotated_lhs = SSA_RE.sub(_add_count, lhs)
-        # Reassemble, preserving spacing and comments
-        new_line = f"{annotated_lhs}={rhs}{comment}"
-        out_lines.append(new_line)
-    # Preserve trailing newline behavior similar to input
-    return "\n".join(out_lines) + ("\n" if mlir_text.endswith("\n") else "")
-
-
-###########################################################
-# Optionally annotate each use of a constant by its value.
-
-DENSE_PAYLOAD_RE = re.compile(r"dense<([^>]+)>\s*:\s*tensor<([^>]+)>", re.IGNORECASE)
-
-
-def is_rank0_or_size1_tensor(type_inner: str) -> bool:
-    """
-    Given the inner part of tensor<...> (i.e., what's inside the angle brackets),
-    return True if it is a rank-0 tensor (just a type like 'f32'/'i64') or a
-    rank-1 size-1 tensor (e.g., '1xf32', '1xi64').
-    """
-    inner = type_inner.strip()
-    # Rank-0: just a scalar type token, no 'x'
-    if "x" not in inner and "[" not in inner and "]" not in inner:
-        return True
-    # Rank-1 size-1: '1x<type>'
-    m = re.match(r"^\s*1x([A-Za-z0-9_<>:?$\-\[\]]+)\s*$", inner)
-    return m is not None
-
-
-def extract_onnx_dense_scalar(rhs: str) -> str | None:
-    """
-    Return the scalar-like value string if RHS contains:
-      onnx.Constant ... dense<...> : tensor<...>
-    and the tensor type is rank-0 or rank-1 size-1.
-    Otherwise return None.
-    """
-    m = DENSE_PAYLOAD_RE.search(rhs)
-    if not m:
-        return None
-    payload = m.group(1).strip()  # e.g., '64', '1.0', '[1, -1, 12, 64]'
-    tensor_inner = m.group(2).strip()  # e.g., '1xi64', 'f32', '4xi64'
-    if not is_rank0_or_size1_tensor(tensor_inner):
-        return None
-    # Reject vector/array payloads like '[1, -1, 12, 64]'
-    if payload.startswith("[") and payload.endswith("]"):
-        return None
-    return payload
-
-
-def collect_onnx_dense_scalar_constants(mlir_text: str) -> dict[str, str]:
-    """
-    Collect { '%name': '<value>' } for onnx.Constant with dense<...> payloads
-    when the type is rank-0 or size-1 tensor.
-    """
-    const_vals: dict[str, str] = {}
-    for line in mlir_text.splitlines():
-        lhs, rhs = split_lhs_rhs_strip_comment(line)
-        if not lhs:
-            continue
-        # Quick pre filter for 'onnx.Constant' and 'dense<'
-        if "onnx.Constant" not in rhs or "dense<" not in rhs:
-            continue
-        value = extract_onnx_dense_scalar(rhs)
-        if value is None:
-            continue
-        for name in SSA_RE.findall(lhs):
-            const_vals[name] = value
-    return const_vals
-
-
-def _annotate_scalar_constant_uses_in_code_segment(
-    code: str, const_vals: dict[str, str]
-) -> str:
-    """Append '/<value>' to each SSA *use* that is a scalar constant."""
-
-    def repl(m: re.Match) -> str:
-        name = m.group(0)
-        if name in const_vals:
-            return f"{name}={const_vals[name]}"
-        return name
-
-    return SSA_RE.sub(repl, code)
-
-
-def annotate_constant_uses_with_values(mlir_text: str) -> str:
-    """
-    Annotate each *use* of a constant discovered by collect_onnx_dense_scalar_constants
-    by appending '/<value>' after the SSA name.
-
-    - LHS (definitions) are not modified.
-    - RHS (and lines without '=') get '%c' -> '%c/<value>' for constants.
-    - Comments are preserved.
-    """
-    const_vals = collect_onnx_dense_scalar_constants(mlir_text)
-
-    out_lines = []
-    for line in mlir_text.splitlines():
-        code, comment = split_comment(line)
-
-        if "=" in code:
-            lhs, rhs = code.split("=", 1)
-            # Do NOT change LHS (definitions); only annotate RHS uses
-            annotated_rhs = _annotate_scalar_constant_uses_in_code_segment(
-                rhs, const_vals
-            )
-            new_line = f"{lhs}={annotated_rhs}{comment}"
-            out_lines.append(new_line)
-        else:
-            # No definition; annotate whole line for constant uses
-            annotated = (
-                _annotate_scalar_constant_uses_in_code_segment(code, const_vals)
-                + comment
-            )
-            out_lines.append(annotated)
-
-    # Preserve trailing newline behavior similar to input
-    return "\n".join(out_lines) + ("\n" if mlir_text.endswith("\n") else "")
-
-
-###########################################################
 # Print the pass that the user is interested in.
 
 
@@ -456,10 +266,13 @@ def print_pass(filename, id, after, pass_name=None):
     message += f' from file "{filename}".'
     # Actual printing
     mlir_text = "".join(pass_listing[i] for i in range(start_id, end_id + 1))
+    # -w must run before -c: -c appends "/N" and "=value" suffixes onto SSA
+    # names, which then fail the head regexes -w's dispatchers use to
+    # recognize generic ops / returns, silently skipping their wrap.
+    if args.wrap:
+        mlir_text = apply_wrap(mlir_text)
     if args.comments:
-        use_counts = count_output_uses(mlir_text)
-        mlir_text = annotate_mlir_with_use_counts(mlir_text, use_counts)
-        mlir_text = annotate_constant_uses_with_values(mlir_text)
+        mlir_text = apply_comments(mlir_text)
     print(f"{message}\n\n", mlir_text, f"\n\n{message}")
